refac: move remoteclient api manager to a separate module
